Support AllToAllTokenDispatcher
